fix type data

+++ koiauction/src/main/java/fall24/swp391/g1se1868/koiauction/service/UserService.java
@@ -157,7 +157,7 @@ public class UserService {
     }
 
     @Transactional
-    public User updateUserRole(Long id, String role) {
+    public User updateUserRole(Integer id, String role) {
         Optional<User> userOptional = userRepository.findById(id);
         if (userOptional.isPresent()) {
             User user = userOptional.get();
